Brodie Commit moved functions into menu.c cleaned main fixed scan function
