+++ PolygonChecker/main.c
@@ -45,13 +45,6 @@ int main() {
 			
 			fillPointsArray(pointsArray);
 
-							/// just to test points ///
-			printf("point1: X = %lf  Y = %lf\n", pointsArray[0].x, pointsArray[0].y);
-			printf("point2: X = %lf  Y = %lf\n", pointsArray[1].x, pointsArray[1].y);
-			printf("point3: X = %lf  Y = %lf\n", pointsArray[2].x, pointsArray[2].y);
-			printf("point4: X = %lf  Y = %lf\n", pointsArray[3].x, pointsArray[3].y);
-							/// just to test points ///
-
 			printf("Perimeter: %lf\n", findPerimeter(pointsArray));
 			bool rectangleCheck = analyze4Points(&pointsArray);
 			if (rectangleCheck != true) {
@@ -59,7 +52,7 @@ int main() {
 			}
 			else {
 				printf("This is a rectangle\n");
-				//findArea(pointsArray);
+				getArea(pointsArray);
 			}
 
 			// free points after case finished
@@ -78,55 +71,9 @@ int main() {
 	return 0;
 }
 
-void printWelcome() {
-	printf("\n");
-	printf(" **********************\n");
-	printf("**     Welcome to     **\n");
-	printf("**   Polygon Checker  **\n");
-	printf(" **********************\n");
-}
 
 
-int printShapeMenu() {
-	printf("1. Triangle\n");
-	printf("2. Rectangle\n");
-	printf("0. Exit\n");
-
-	int shapeChoice;
-	bool validity = false;
-	do {
-		printf("Enter number: ");
-		if (scanf("%1d", &shapeChoice) != 1)
-		 {
-			while (getchar() != '\n');
-			printf("Invalid input\n");
-		}
-
-		if (shapeChoice == 1 || shapeChoice == 2 || shapeChoice == 0) {
-			validity = true;
-		}
-		else {
-			printf("Invalid input\n");
-			while (getchar() != '\n');
-		}
-
-	} while (validity == false);
-	return shapeChoice;
-}
 
 
-double* getTriangleSides(double* triangleSides) {
-	int sideNum[TRIANGLESIDES] = { 1, 2, 3 };
-	printf("Enter the three sides of the triangle: \n");
-	for (int i = 0; i < TRIANGLESIDES; i++)
-	{
-		printf("side%d: ", sideNum[i]);
-		if (scanf("%lf", &triangleSides[i]) != 1) {
-			printf("Invalid input. Please enter number\n");
-			i--;
 
-			while (getchar() != '\n');
-		}
-	}
-	return triangleSides;
-}
+
