+++ PolygonChecker/menu.h
@@ -1,1 +1,5 @@
 #pragma once
+
+void printWelcome();
+
+int printShapeMenu();
