+++ PolygonChecker/triangleSolver.c
@@ -1,8 +1,27 @@
+#define _CRT_SECURE_NO_WARNINGS
+#include "main.h"
 #include <stdio.h>
 #include <stdbool.h>
 #include <math.h>
 
 #include "triangleSolver.h"
+
+
+double* getTriangleSides(double* triangleSides) {
+	int sideNum[TRIANGLESIDES] = { 1, 2, 3 };
+	printf("Enter the three sides of the triangle: \n");
+	for (int i = 0; i < TRIANGLESIDES; i++)
+	{
+		printf("side%d: ", sideNum[i]);
+		if (scanf("%lf", &triangleSides[i]) != 1) {
+			printf("Invalid input. Please enter number\n");
+			i--;
+
+			while (getchar() != '\n');
+		}
+	}
+	return triangleSides;
+}
 
 char* analyzeTriangle(double side1, double side2, double side3) {
 	char* result = "Not a Triangle";
@@ -30,3 +49,4 @@ double getAngle(double a, double b, double c) {
 	// formula for findng angle A with sides a b c. Also, it convers rad to deg
 	return (180/M_PI * acos( ((b * b) + (c * c) - (a * a)) / (2 * b * c) ));
 }
+
