+++ PolygonChecker/triangleSolver.h
@@ -2,6 +2,8 @@
 
 #define M_PI 3.14159265358979323846
 
+double* getTriangleSides(double* triangleSides);
+
 char* analyzeTriangle(double side1, double side2, double side3);
 
 // Takes three side lengths of a triangle,
